GitButler Integration Commit
This is an integration commit for the virtual branches that GitButler is tracking.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - Add routes and html page (refs/gitbutler/add-routes-and-html-page)
   branch head: 65ccc9c808beb69d5d03b2bbdaaf721654444d7d
   - index.js
   - layouts/dispenser.html
   - public/index.html

Your previous branch was: refs/heads/main

The sha for that commit was: c2d20bec0310114ddaad707a21ce71711142cef4

For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch